+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-﻿### 4.1.88 (2022-05-18 20:24Z)
+﻿### 4.1.89 (2022-05-19 20:34Z)
+  * AWS Tools for PowerShell now use AWS .NET SDK 3.7.278.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
+  * Amazon Lookout for Metrics
+    * Modified cmdlet New-LOMAlert: added parameter SNSConfiguration_SnsFormat.
+
+### 4.1.88 (2022-05-18 20:24Z)
   * AWS Tools for PowerShell now use AWS .NET SDK 3.7.277.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
   * Amazon App Mesh
     * Modified cmdlet New-AMSHMesh: added parameter ServiceDiscovery_IpPreference.